+++ presentation/src/main/java/way/presentation/schedule/validates/DeleteScheduleValidator.java
@@ -4,11 +4,11 @@ import org.springframework.stereotype.Component;
 
 import way.application.utils.exception.BadRequestException;
 import way.application.utils.exception.ErrorResult;
-import way.presentation.schedule.vo.req.DeleteScheduleRequest;
+import way.presentation.schedule.vo.req.ScheduleRequestVo;
 
 @Component
 public class DeleteScheduleValidator {
-	public void validate(DeleteScheduleRequest request) {
+	public void validate(ScheduleRequestVo.DeleteScheduleRequest request) {
 		validateScheduleSeq(request.scheduleSeq());
 		validateMemberSeq(request.creatorSeq());
 	}
