+++ presentation/src/main/java/way/presentation/schedule/validates/ModifyScheduleValidator.java
@@ -6,11 +6,11 @@ import org.springframework.stereotype.Component;
 
 import way.application.utils.exception.BadRequestException;
 import way.application.utils.exception.ErrorResult;
-import way.presentation.schedule.vo.req.ModifyScheduleRequest;
+import way.presentation.schedule.vo.req.ScheduleRequestVo;
 
 @Component
 public class ModifyScheduleValidator {
-	public void validate(ModifyScheduleRequest request) {
+	public void validate(ScheduleRequestVo.ModifyScheduleRequest request) {
 		validateScheduleSeq(request.scheduleSeq());
 		validateTitle(request.title());
 		validateStartTime(request.startTime());
