--- presentation/src/main/java/way/presentation/schedule/vo/req/DeleteScheduleRequest.java
@@ -1,15 +1,0 @@
-package way.presentation.schedule.vo.req;
-
-import way.application.service.schedule.dto.request.DeleteScheduleRequestDto;
-
-public record DeleteScheduleRequest(
-	Long scheduleSeq,
-	Long creatorSeq
-) {
-	public DeleteScheduleRequestDto toDeleteScheduleRequestDto() {
-		return new DeleteScheduleRequestDto(
-			this.scheduleSeq,
-			this.creatorSeq
-		);
-	}
-}
