--- presentation/src/main/java/way/presentation/schedule/vo/res/ModifyScheduleResponse.java
@@ -1,6 +1,0 @@
-package way.presentation.schedule.vo.res;
-
-public record ModifyScheduleResponse(
-	Long scheduleSeq
-) {
-}
