--- presentation/src/main/java/way/presentation/schedule/vo/res/SaveScheduleResponse.java
@@ -1,7 +1,0 @@
-package way.presentation.schedule.vo.res;
-
-public record SaveScheduleResponse(
-	Long scheduleSeq
-) {
-
-}
